[core] Warning fixes

+++ 3party/agg/agg_color_rgba.h
@@ -323,13 +323,13 @@ namespace agg
         //--------------------------------------------------------------------
         static AGG_INLINE double to_double(value_type a)
         {
-            return double(a) / base_mask;
+            return double(a) / double(base_mask);
         }
 
         //--------------------------------------------------------------------
         static AGG_INLINE value_type from_double(double a)
         {
-            return value_type(uround(a * base_mask));
+            return value_type(uround(a * double(base_mask)));
         }
 
         //--------------------------------------------------------------------
@@ -701,13 +701,13 @@ namespace agg
         //--------------------------------------------------------------------
         static AGG_INLINE double to_double(value_type a)
         {
-            return double(a) / base_mask;
+            return double(a) / double(base_mask);
         }
 
         //--------------------------------------------------------------------
         static AGG_INLINE value_type from_double(double a)
         {
-            return value_type(uround(a * base_mask));
+            return value_type(uround(a * double(base_mask)));
         }
 
         //--------------------------------------------------------------------
@@ -888,7 +888,7 @@ namespace agg
         AGG_INLINE self_type gradient(const self_type& c, double k) const
         {
             self_type ret;
-            calc_type ik = uround(k * base_mask);
+            calc_type ik = uround(k * double(base_mask));
             ret.r = lerp(r, c.r, ik);
             ret.g = lerp(g, c.g, ik);
             ret.b = lerp(b, c.b, ik);
@@ -1120,7 +1120,7 @@ namespace agg
         //--------------------------------------------------------------------
         static AGG_INLINE value_type mult_cover(value_type a, cover_type b) 
         {
-            return value_type(a * b / cover_mask);
+            return value_type(a * double(b) / double(cover_mask));
         }
 
         //--------------------------------------------------------------------
